+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/ExprValueExtensions.kt
@@ -14,6 +14,7 @@
 
 package org.partiql.lang.eval
 
+import com.amazon.ion.Decimal
 import com.amazon.ion.IntegerSize
 import com.amazon.ion.IonInt
 import com.amazon.ion.IonStruct
@@ -22,6 +23,11 @@ import com.amazon.ion.IonType
 import com.amazon.ion.IonValue
 import com.amazon.ion.Timestamp
 import com.amazon.ion.system.IonSystemBuilder
+import com.amazon.ionelement.api.field
+import com.amazon.ionelement.api.ionDecimal
+import com.amazon.ionelement.api.ionInt
+import com.amazon.ionelement.api.ionStructOf
+import com.amazon.ionelement.api.toIonValue
 import org.partiql.lang.ast.SourceLocationMeta
 import org.partiql.lang.errors.ErrorCode
 import org.partiql.lang.errors.Property
@@ -57,6 +63,8 @@ import org.partiql.types.StringType
 import org.partiql.types.SymbolType
 import org.partiql.types.TimeType
 import org.partiql.types.TimestampType
+import org.partiql.value.datetime.TimestampWithTimeZone
+import org.partiql.value.datetime.TimestampWithoutTimeZone
 import java.math.BigDecimal
 import java.math.MathContext
 import java.math.RoundingMode
@@ -74,6 +82,7 @@ const val BAG_ANNOTATION = "\$bag"
 const val DATE_ANNOTATION = "\$date"
 const val TIME_ANNOTATION = "\$time"
 const val GRAPH_ANNOTATION = "\$graph"
+const val TIMESTAMP_ANNOTATION = "\$timestamp"
 
 /**
  * Wraps the given [ExprValue] with a delegate that provides the [OrderedBindNames] facet.
@@ -135,9 +144,12 @@ fun ExprValue.dateValue(): LocalDate =
 fun ExprValue.timeValue(): Time =
     scalar.timeValue() ?: errNoContext("Expected time: $this", errorCode = ErrorCode.EVALUATOR_UNEXPECTED_VALUE_TYPE, internal = false)
 
+@Deprecated("ExprValue.timestampValue is deprecated", replaceWith = ReplaceWith("partiQLTimestampValue()"))
 fun ExprValue.timestampValue(): Timestamp =
     scalar.timestampValue() ?: errNoContext("Expected timestamp: $this", errorCode = ErrorCode.EVALUATOR_UNEXPECTED_VALUE_TYPE, internal = false)
 
+fun ExprValue.partiQLTimestampValue(): org.partiql.value.datetime.Timestamp =
+    scalar.partiQLTimestampValue() ?: errNoContext("Expected timestamp: $this", errorCode = ErrorCode.EVALUATOR_UNEXPECTED_VALUE_TYPE, internal = false)
 fun ExprValue.stringValue(): String =
     scalar.stringValue() ?: errNoContext("Expected string: $this", errorCode = ErrorCode.EVALUATOR_UNEXPECTED_VALUE_TYPE, internal = false)
 
@@ -736,7 +748,7 @@ fun ExprValue.toIonValue(ion: IonSystem): IonValue =
                 addTypeAnnotation(DATE_ANNOTATION)
             }
         }
-        ExprValueType.TIMESTAMP -> ion.newTimestamp(timestampValue())
+        ExprValueType.TIMESTAMP -> partiQLTimestampValue().toIonValue(ion)
         ExprValueType.TIME -> timeValue().toIonValue(ion)
         ExprValueType.SYMBOL -> ion.newSymbol(stringValue())
         ExprValueType.STRING -> ion.newString(stringValue())
@@ -777,3 +789,19 @@ private fun ExprValue.toIonStruct(ion: IonSystem): IonStruct {
         }
     }
 }
+
+private fun org.partiql.value.datetime.Timestamp.toIonValue(ion: IonSystem) =
+    when (this) {
+        is TimestampWithTimeZone -> ion.newTimestamp(this.ionTimestampValue)
+        is TimestampWithoutTimeZone -> {
+            val ts = this@toIonValue
+            ionStructOf(
+                field("year", ionInt(ts.year.toLong())),
+                field("month", ionInt(ts.month.toLong())),
+                field("day", ionInt(ts.day.toLong())),
+                field("hour", ionInt(ts.hour.toLong())),
+                field("minute", ionInt(ts.minute.toLong())),
+                field("second", ionDecimal(Decimal.valueOf(ts.decimalSecond)))
+            ).withAnnotations(TIMESTAMP_ANNOTATION).toIonValue(ion)
+        }
+    }